pyproject.toml
[project]
name = "intra-file-share"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.6",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "httpx>=0.28.1",
    "httpx-oauth>=0.16.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "sentry-sdk>=2.56.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.40.0",
]

[tool.uv.workspace]
members = [
"intra_file_share"
]
